fixed some warnings

+++ Mixpanel/MPABTestDesignerSnapshotResponseMessage.m
@@ -53,7 +53,7 @@
 - (NSString *)getImageHash:(NSData *)imageData
 {
     unsigned char result[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(imageData.bytes, imageData.length, result);
+    CC_MD5(imageData.bytes, (uint)imageData.length, result);
     NSString *imageHash = [NSString stringWithFormat:@"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
                            result[0], result[1], result[2], result[3],
                            result[4], result[5], result[6], result[7],
